pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "bharatpe-payment-plugin"
version = "1.0.0"
description = "Drop-in BharatPe UPI payment plugin for python-telegram-bot"
readme = "PLUGIN_README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["bharatpe", "upi", "telegram", "bot", "payment", "india"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "python-telegram-bot>=20.0",
    "requests>=2.28",
    "psycopg2-binary>=2.9",
    "Pillow>=9.0",
    "qrcode[pil]>=7.0",
]

[project.urls]
Homepage = "https://github.com/pay0bot/bharatpe-payment-plugin"
Documentation = "https://github.com/pay0bot/bharatpe-payment-plugin/blob/main/INTEGRATION.md"
"Bug Tracker" = "https://github.com/pay0bot/bharatpe-payment-plugin/issues"

[tool.setuptools.packages.find]
include = ["payment_plugin*"]
